MAJOR: cgi/livestatus: Index custom variables in a JSON sub-object
Tagged as MAJOR because it changes the JSON format; however that does not
break kompot-wui because it uses CV_COLUMNS mode where an explicit set
of custom variables are returned in host/service objects.

Here we introduce the ability to define a per-property formating option.
This is done in the json_property() function. Most of the formating
logic gets out of the main { ... } awk section.