MAJOR: cgi/livestatus: Support both host and services custom variables
This patch makes a change when CV_COLUMNS mode is enabled, which is
the default and what is used with kompot-wui. It is applied here to
the _TAGS variable where we want to merge tags from both host and
services custom variables.

The change is in function computed_columns(), which does the CV_COLUMNS
processing to add new columns from custom variables.

Also more characters gets permitted in search, particularly to include
the hash (#) that makes sense as a prefix for a tag.