BUG/MEDIUM: addons/rest: Quote macro arguments in commands
A patch that expands non-existing macros to an empty string was merged
in Nagios 4.5.7. This is a breaking change. Before that an undefined macro
like $_SERVICE_SOMME_VAR$ would result in $ as argument in shell exec and
check_rest treat that value specially. Now the expansion results in an
empty string and check_rest fails to parse arguments, hence the need to
quote, which should have been done the first time.