MINOR: update-nagios: Improve display of errors on apply failure
For docker installations, this goes along with commit bb614587 "MINOR:
Update nagios runit to verify configuration on start".

For now, we stick with the current implementation that attempts to
extract new configuration errors from logs, as opposed to run
nagios -v from the update-nagios script because this script does not
necessarily know which main nagios.cfg to use or which user to run as.

Before docker support was added, last error logs were retrieved with
journalctl. This worked because nagios unit file has a ExecStartPre to
run nagios -v in order to check configuration, that output gets eaten
by journal/syslog. When docker support was added I made a change to
look in /var/log/nagios/nagios.log instead; it broke errors display
on non-docker environment.

This commits changes erorrs logs extraction to get back to a
journal/syslog source. This was tested both on docker and classic
installations.